Time: 8 ms (14.4%), Space: 103.8 MB (3.11%) - LeetHub

+++ 3689-maximum-total-subarray-value-i/3689-maximum-total-subarray-value-i.cpp
@@ -3,6 +3,6 @@ class Solution {
 public:
     long long maxTotalValue(vector<int>& nums, int k) {
         ll ma = *max_element(nums.begin(),nums.end()), mi = *min_element(nums.begin(),nums.end());
-        return (ll)k*(ma-mi);
+        return k*(ma-mi);
     }
 };
